chore: move group components to own folder
